HW-4 SharedPreferenceUtil get/set 만 하도록 수정
