+++ class15/hhi-5258/HHiApplication/app/src/main/java/com/hhi/myapplication/data/local/NaverLocalDataSourceImpl.kt
@@ -1,11 +1,33 @@
 package com.hhi.myapplication.data.local
 
 import com.hhi.myapplication.App
+import org.json.JSONArray
+
+private const val PREF_QUERY_LIST: String = "pref_query_list"
 
 class NaverLocalDataSourceImpl : NaverLocalDataSource {
     override fun saveQuery(query: String) {
-        App.prefs.saveQuery(query)
+        val queryList = getQueryList().toMutableList()
+        queryList.add(query)
+
+        if (queryList.size > 5) {
+            queryList.removeAt(0)
+        }
+
+        App.prefs.setString(JSONArray(queryList).toString(), PREF_QUERY_LIST)
     }
 
-    override fun getQueryList(): List<String> = App.prefs.getQueryList()
+    override fun getQueryList(): List<String> {
+        val queryListJSONString = App.prefs.getString(PREF_QUERY_LIST)
+        val queryList = mutableListOf<String>()
+
+        queryListJSONString?.let {
+            val jsonArray = JSONArray(it)
+            for (i in 0 until jsonArray.length()) {
+                queryList.add(jsonArray[i].toString())
+            }
+        }
+
+        return queryList
+    }
 }
